Went back to the old HFSM code and made some adjustments.
Transitions no longer have to be held in the state itself, and can
be triggered at any time, not just during updates. This will let us
tie transitions to events much more easily. The new/old HFSM code was
starting to get a lot more complicated than it was worth, so I'm
probably going to take it out to work on it separately.