Tweaked physics, added Mouse-Over Glow, and centered baskets.
Strawberries now have much more drag, but gravity has also been scaled
up to match. So they'll still fall the same way, but won't drift as far
from the mouse cursor if they collide with something while being
dragged.

The glow effect helps you know which berry you're selecting with the
mouse. It also has been helping me find a bug with the object spawner
where the first set of strawberries won't respond to input (these won't
glow when you mouse over them; the working berries do).

I'm starting to think that picking berries one-at-a-time is too much of
a hassle as well as unrealistic, so next I'm going to try to find a way
for the player to pick up a basket to take with them into the pick
state.