Added left hand container, experimenting with state machines again.
I'm thinking about giving the strawberry objects their own state
machines so that we can tell when they're in the field, being dragged,
held in the player's other hand, in a basket, etc. and give them their
own behavior in each. Tried messing around with the reactive HFSM again
but I'm going to settle with my first version, which seems to be the
most reliable.