Strawberries now change color depending on their quality.
I added a very simple custom shader that changes the color of the
strawberry depending on it's in-game quality value. The shader should
support a texture once Juan comes up with one, but it can always be
tweaked in case it doesn't. However, I'm getting a slowdown on my
desktop computer, which is strange, since all my added code should only
apply when the strawberry is created. I'll keep investigating.