Mostly code cleanup.
GameStateManager now stores its states in a dictionary like
StrawberryStateMachine does. Also did some cleanup on
StrawberryComponent, hoping to move some of the drag functionality to
the state system. It's still in-progress, so for now you can grab
strawberries from any state, but this will be fixed in the next update.