Did some fairly radical restructuring in preparation for adding a menu system. There's a new StateMachine component that handles states and transitions (was using internal dictionaries for those before - bad idea in retrospect). The strawberry state system has been simplified a bit, since I got rid of a few redundant states.
Also switched to a new mouse-drag component which is a bit more accurate and integrated with the state system. While strawberries don't rotate with the camera anymore (which I liked), it should cut down on the "drift" problem some people have had. I'll work on getting rotation working again at some point so we can have the best of both worlds.