Strawberry bushes now have leaves again.
The leaves are now arranged around the center of the ground cell, so now
they even look like strawberry bushes instead of clovers. Had a random
freeze-on-startup bug which went away after doing some state system
tweaks. I don't know where the problem was, but I'm keeping an eye on it
in case it comes up again.