Fixed the long startup time.
Turns out it was because strawberries were spawning right on top of the
baskets, causing them to get stuck in a loop when the deposit transition
triggered when it shouldn't. This could probably be fixed more
thoroughly by making state system adjustments, but for now I fixed it by
changing the spawn location to be far away from the player.