Reworked how strawberries are sized. Now ripe strawberries tend to be slightly bigger than unripe ones, though there is still plenty of variation.
Also working on changing how the mouseover glow works. Instead of using Unity's built-in glow effect, I'm going to try using a billboard object. That way I can make the glow scale with the strawberry and appear on top of other objects like leaves. I can also make it so that the glow only appears when the strawberry can actually be picked up.