Got started on UI changes.
The compass arrows now use Unity's own UI system instead of me trying to
hack it with a game object. For starters, it means the compass will be
visible on top of other objects. It's also clickable now, which I used
to add mouse controls since I noticed that the playtesters often clicked
the arrows, expecting them to do something. There's also a nice fade-
in/out effect on the arrows when you mouse over them.