Main menu connects to the game scene
Also experimented with different ways of setting up the sky. I think we
should stick with cylinders but for some reason I'm getting a framerate
drop with the new one I made. I'll see if it's just a problem with my
machine (remember that the laptops at the NEST are probably better than
anything I'm running, so as long as it's basically playable on mine, it
should be okay there).