Sundry fixes in preparation for adding an endgame state.
Fixed the bee model to use multiple planes instead of a single one so
that it doesn't disappear when looking at it sideways. As a bonus, I
also downloaded a couple of double-sided shaders which are probably
better than the homerolled version I made. I also consolidated the UI
visibility component into the regular one and added the ability to chain
them together. Fixed some UI components to fit better, as well.