Switched the Pace Manager to a slider instead of a state system.
I've found that this gives you a lot more control over movement and
makes it explicit when you're moving while packing, which used to just
be an undocumented feature of the "fast" pace. Also added key controls
so you can speed up, slow down, or instantly stop while still holding
onto a strawberry (E,Q, and Space, or +,-, and 0 on the numpad).