Displacers now affect collision meshes.
Also added a new displacer that lets you reduce displacement near the
edges of a mesh. I used it on the ground mesh to fade it to 0 when it
gets near the edge between the generated terrain and the flat plains
making up the ground on the other rows, so now they should be seamless
even when going through a break.