Time system changes, new cart texture, and fixes to scoring and ground.
Game timer can now display in game time as well as real time (so we can
show the time as 6:00 PM instead of 0 seconds). Also tweaked the cart
texture and added a diffuse+normal map, so now it looks more like a
wireframe, though it's still a bit chunky-looking. Fixed a few bugs with
the basket UI, and found out that the displacement axis on the plant
rows was wrong, so now they should hug the ground a bit more closely.