Asset cleanup and a new strawberry model/texture.
Moved a lot of the scripts that I had left in Assets into their proper
folders. Also made a new strawberry model using the texture we got from
Juan's brother. It works in the editor, but I ran into a strange bug in
gameplay with Unity and normal maps where the normal map won't show up
on the strawberry unless the material is selected in the editor. Since
the normal map is set and shows up immediately after viewing the
material, I think it's a bug with Unity itself and not something I can
fix. Maybe updating Unity to the latest version will fix it. Since I was
thinking of updating to the beta in order to get Unity Analytics, I
might try updating in the next commit, but that might break a few
things. We'll see.