Added a new Displace modifier.
Now displacement can affect an object's in-engine position in the game
as well as its vertices. I'm currently using this for attaching the
strawberry leaves to the ground (fixing the floating stems problem), but
it could also be used for strawberries, or to make the cart/camera
follow the ground.

Earlier I tried optimizing the leaf generation system, since that seemed
to be a performance bottleneck. I ended up reverting it, unfortunately,
but I'll try tackling it again tomorrow when I'm not overtired.