New optimization system for strawberry leaves.
Instead of drawing every time a new ground cell is made (which caused a
ton of stutter), leaves are gradually added to the scene or recycled
from distant cells as the player moves. This means it takes a few
seconds for the scene to fill up with leaves, but it cuts down the
geometry that the computer has to draw and streamlines movement. It
could still be better (for instance, adding billboards for really far
away plants), but it's a huge improvement what we had before.

However, I think I might have broken strawberry respawning in the
process. I'll try to take a look at it and fix it in the next update.
I'm just committing what I have since at least the leaf system seems to
be working.