Tweaked the cart material and added a credits screen.
The cart now looks more like metal than the default material I was using
earlier. Still no texture, but with such a thin frame I don't know if we
really need one. Also, the credits screen is technically in but needs
material. I'll try to ask Dvera for the names needed to fill it out.