Got most of the settings menu internationalized.
Also improved the ObjectVisibility component so it doesn't update on
every frame. It was a bit buggy at first but it seems to be working fine
now.

At some point I might want to move the ground cells over to a state
system like the leaves, because there's still a bit of a stutter every
time the player moves forward. If it's not that big a problem I might
just leave it alone, though.