Ground model fixes.
I couldn't find the original file that had the improved ground model in
it, so I had to reconstruct it. It looks good now, but I'm a little
disappointed I had to go over old territory. Hopefully this will be the
last fix and I can get back to adding features.