Minor tweaks
Ground tiles no longer cast their own shadows for performance reasons
(needs testing). Reworked the footsteps component a bit so that the
rhythm can be tweaked independently of the player's speed. Added two
buttons to the loading screen. One lets the player return to the menu
before the game loads in case they clicked "start" by accident. The
other will eventually bring up a copy of the controls menu so the player
can review if they didn't see it earlier.