Redid the sky cylinder and adjusted footstep rhythm.
The new panorama uses the entire texture instead of just part of a
photo, making it easier in case I want to use a different texture. Also
made the footsteps rhythm a bit better and tried (again) to fix the
"get_seed can only be called from the main thread" problem (and
thankfully UniRx seems to have a way to call things from the main thread
easily).