Mostly art tweaks, but some coding as well.
Added Juan's key sprites and cover image. Added labels to most of the
control menu icons. Put a wood texture on the button sprite. Removed the
bee from the cart (don't worry, it's still in Assets so we can put it
back if we want). Added a "cheat sheet" to the cart that tells the
player what ranges of strawberry ripeness are acceptable. I got the
suggestion from the Unity meetup earlier this week, and it makes sense
for newer players who might not know what they're doing yet.