Fixed basket icons and made the entire basket system reactive.
Now the new BasketIcon component can be used for 2d and 3d
interchangeably. In addition, this fixed the bug where the overflow icon
would only appear once you were within the target weight: now it appears
whenever you have a strawberry out of bounds.

There's still a bug where the game records more baskets than are
actually in the game on repeated sessions. I'll try to fix it in the
next update.