Strawberries can no longer be "pushed" through objects.
When you pick up a strawberry, the physics engine now does a raycast to
find the closest collidable object to you and brings the strawberry in
front of it if it was previously behind it. This fixes several problems
where the player would accidentally drop strawberries out of the basket
or bury them in the floor.