Added "slide" buttons
These let you make quick adjustments to your position when you're in
Pick or Pack mode, so you don't have to set your pace and quickly go in-
and-out of Look mode just to pick up a strawberry that's barely out of
reach. Also added gradiented versions of Juan's forward-and-back
symbols.