Mostly interface fixes.
Strawberry glows and footprints now fade in gradually instead of popping
in and out. Tried to get basket icons to fade out when you mouse over
them, but it caused conflicts with picking up strawberries. I might go
back and fix that later on.

The compass is now scalable, and the pace slider now has an auto-layout
instead of manual placement. In case we ever wanted to make a tablet
version of our game, I figured I should clean up the interface a bit to
make it easier to scale up.