Added a quick select box for settings selection.
It's actually been there for a while, but I never got around to
implementing it until now. The idea is to make it as easy to pick a
settings file as it is to pick a difficulty in a non-computer game. As a
proof-of-concept, I made a set of "quick" rules in the game itself and
saved it out, which works pretty nicely.