Implemented a few minor features, mostly with scores.
You can now sort high scores by date, strawberries, or baskets, though
you need to reload the scores file in order to see it (tried getting it
to work instantly, but no luck so far). Also got total weight, average
weight, and average ripeness working.

As a side-effect, however, it looks like the game is starting with quick
settings instead of the defaults. This is probably just because it's the
first file that C# finds for some reason. I'll try to get the code that
selects the default file for you working again.