Hid the gameplay settings in a separate screen.
That way we can later add player-specific settings like sound volume and
such that are more like what you'd see in a regular game, instead of
overwhelming them with gameplay settings.

Also did a bit more testing on the ObjectVisibility component, and I
discovered that one of the serialization errors I was getting in the
editor was coming from the game not loading completely. If you wait for
the field to load before pressing "Go Back," there's no error. Dunno how
to fix that, but it might just be a Unity problem.