Replaced the timer and added a distance meter.
The new, Pikmin-inspired timer uses icons to better represent how long
the player has until the end of the game. This makes the game more of a
race against the clock which should make it a bit more exciting.

There's currently a soft distance limit of 500 meters but it's not
reflected in scores just yet; right now it just makes the cart icon loop
back around when the player passes it.

Also shortened the default game from 40 minutes to 30, and lowered the
minimum strawberry size so that there are more small "filler" berries to
round out the bigger ones. Density is also slightly increased, so it
doesn't take as many berries to fill up a basket.