More work on the language system.
The old language component now replaces itself with the new version
dynamically whenever possible.

I tried to get this working in the editor so that I wouldn't have to
replace each component manually, but since that's not working I guess
I'll just have to sit down sometime and do that. Also, dynamic text
still uses the old system so I'll have to update that as well. Still,
the new system seems to be handling the load pretty well.

I also changed the formatting on the language files from ANSI to UTF-8,
letting us use accents again.