Reworked Win Conditions.
Now instead of min and max values, scores are determined by how close a
value (like basket weight or berry ripeness) is to a given target. This
gives us the ability to determine how "close" the player is to the
target, giving us a finer-grain way of determining score rather than the
booleans we were using before. The best part is that the class can be
reused for things like the time and distance limits.