Fixed a bug where settings wouldn't update properly.
I added back the apply and revert functions and hooked them up to the UI
again. There's also another GameSettingsComponent layer between the main
settings screen and the advanced screen, so fiddling with the advanced
settings and then cancelling won't clobber the original settings.

Also added a control for the distance target.