Minimized the high score views.
Now that we can calculate a single number from a score, I was able to
cut out a lot of the more intimidating data from the high score screen.
Now it just shows the player's name, date, score, and the basket icons
(with no numbers attached). I kept the old code which I'm going to use
to write the detailed score screen. There's still a bug with the
ScoreTarget that's causing scores to come up negative, but I think that
could just be a math error. Shouldn't be too hard to fix.