Did some subscription cleanup.
Not happy with Unity Editor today. I came across a mass of bugs related
to me not disposing of my subscriptions correctly, which only started
showing up in the final build when you switch from the menu to gameplay
and back. I added some cleanup code and replaced the SettingsForm UI
code with the new UI Adapter. The adapter also keeps track of
subscriptions as they're made and can clean them all up at once, which
is nice. Now I'm suspicious of the rest of my code, though, so I'll have
to do more testing to make sure that it doesn't creep up on us.