Stability fix. Figured out the subscription problem.
I didn't catch it before because most of the observables I was
subscribing to were self-contained within the same behavior as the
subscription. However, the new score, settings, and language frameworks
are more interwoven. I fixed it by searching every instance where I
subscribed to an observable, checked to see if it was self-contained,
and added disposal logic where it wasn't.